:recycle: Refactor: 게시물업로드, 게시물 수정시 공통으로 사용하는 타입 분리(#21)
