+++ src/components/Upload/model.ts
@@ -1,0 +1,11 @@
+interface AccordionData {
+  question: string;
+  answer: string[];
+}
+
+interface AlbumIdData {
+  albumName: string;
+  docId: string;
+}
+
+export type { AccordionData, AlbumIdData };
